Move MockToolWindow to abstractIT class
