+++ src/test/java/network/radicle/jetbrains/radiclejetbrainsplugin/AbstractIT.java
@@ -5,6 +5,7 @@ import com.intellij.notification.Notification;
 import com.intellij.notification.Notifications;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.SystemInfo;
 import com.intellij.openapi.vcs.VcsException;
 import com.intellij.openapi.vcs.changes.Change;
@@ -12,6 +13,7 @@ import com.intellij.testFramework.CoroutineKt;
 import com.intellij.testFramework.HeavyPlatformTestCase;
 import com.intellij.testFramework.PlatformTestUtil;
 import com.intellij.testFramework.ServiceContainerUtil;
+import com.intellij.toolWindow.ToolWindowHeadlessManagerImpl;
 import com.intellij.util.messages.MessageBusConnection;
 import git4idea.GitCommit;
 import git4idea.config.GitConfigUtil;
@@ -193,6 +195,22 @@ public abstract class AbstractIT extends HeavyPlatformTestCase {
 
         @Override
         public void resumeWith(@NotNull Object o) {
+        }
+    }
+
+    public static class MockToolWindow extends ToolWindowHeadlessManagerImpl.MockToolWindow {
+        public MockToolWindow(@NotNull Project project) {
+            super(project);
+        }
+
+        @Override
+        public boolean isAvailable() {
+            return false;
+        }
+
+        @Override
+        public boolean isVisible() {
+            return true;
         }
     }
 }
